refactor: extract ui dialog overlays
